move UInt64 into a namespace
