+++ src/Lib/Elf/Structure/Elf64/Elf64DynamicStructure.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace PhpProfiler\Lib\Elf\Structure\Elf64;
 
-use PhpProfiler\Lib\UInt64;
+use PhpProfiler\Lib\Integer\UInt64;
 
 /**
  * Class Elf64DynamicStructure
